int -> long int

+++ myisfinite.cpp
@@ -8,7 +8,7 @@ int myisfinite(double a)
 {
     int size = sizeof(double) * 8;
         // это из формулы, данной в стандарте:
-    int w = lround(4 * log2(size)) - 13;
+    long int w = lround(4 * log2(size)) - 13;
         //маска на все заполненные биты в E
     uint64_t mask = (1ULL << w) - 1;
 
@@ -24,7 +24,7 @@ int myINForNAN(double a)
 {
     int size = sizeof(double) * 8;
         // это из формулы, данной в стандарте
-    int w = lround(4 * log2(size)) - 13;
+    long int w = lround(4 * log2(size)) - 13;
         //маска на биты мантиссы (если она ноль, то это INF, иначе NAN)
     uint64_t mask = (1ULL << (size - w - 1)) - 1;
         //забавный костыль
